refactor(stream): Split into more files
